Changed: No casting to signed integers for CLZ/CTZ on Windows

+++ include/gaia/external/robin_hood.h
@@ -175,16 +175,16 @@ inline bool CheckEndianess() {
 
 		#pragma intrinsic(ROBIN_HOOD(CTZ))
 		#define ROBIN_HOOD_COUNT_TRAILING_ZEROES(x)                                                                        \
-			[](size_t mask) noexcept -> int {                                                                                \
+			[](size_t mask) noexcept -> size_t {                                                                             \
 				unsigned long index;                                                                                           \
-				return ROBIN_HOOD(CTZ)(&index, mask) ? static_cast<int>(index) : ROBIN_HOOD(BITNESS);                          \
+				return ROBIN_HOOD(CTZ)(&index, mask) ? index : ROBIN_HOOD(BITNESS);                                            \
 			}(x)
 
 		#pragma intrinsic(ROBIN_HOOD(CLZ))
 		#define ROBIN_HOOD_COUNT_LEADING_ZEROES(x)                                                                         \
-			[](size_t mask) noexcept -> int {                                                                                \
+			[](size_t mask) noexcept -> size_t {                                                                             \
 				unsigned long index;                                                                                           \
-				return ROBIN_HOOD(CLZ)(&index, mask) ? static_cast<int>(index) : ROBIN_HOOD(BITNESS);                          \
+				return ROBIN_HOOD(CLZ)(&index, mask) ? index : ROBIN_HOOD(BITNESS);                                            \
 			}(x)
 	#else
 		#if ROBIN_HOOD(BITNESS) == 32
